backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain
langchain-openai==0.1.1
langchain-core
openai<2.0.0,>=1.10.0
judgeval==0.0.50
python-dotenv==1.0.1
pydantic==2.5.0
pypdf2==3.0.1
python-docx==1.1.0